chore(core): divide useEventsStore subscriptions
