+++ packages/sanity/src/core/store/events/getDocumentAtRevision.ts
@@ -1,0 +1,48 @@
+/* eslint-disable no-console */
+/* eslint-disable max-nested-callbacks */
+import {type Observable, of} from 'rxjs'
+import {catchError, map, startWith, tap} from 'rxjs/operators'
+import {type SanityClient, type SanityDocument} from 'sanity'
+
+import {type EventsStoreRevision} from './types'
+
+const documentRevisionCache: Record<string, SanityDocument> = Object.create(null)
+
+export function getDocumentAtRevision({
+  client,
+  documentId,
+  revisionId,
+}: {
+  client: SanityClient
+  documentId: string
+  revisionId: string | null | undefined
+}): Observable<EventsStoreRevision | null> {
+  if (!revisionId) {
+    return of(null)
+  }
+  const cacheKey = `${documentId}@${revisionId}`
+  const dataset = client.config().dataset
+  if (documentRevisionCache[cacheKey]) {
+    return of({document: documentRevisionCache[cacheKey], loading: false, revisionId: revisionId})
+  }
+  return client.observable
+    .request<{documents: SanityDocument[]}>({
+      url: `/data/history/${dataset}/documents/${documentId}?revision=${revisionId}`,
+      tag: 'get-document-revision',
+    })
+    .pipe(
+      map((response) => {
+        const document = response.documents[0]
+        return {document: document, loading: false, revisionId: revisionId}
+      }),
+      tap((resp) => {
+        documentRevisionCache[cacheKey] = resp.document
+      }),
+      catchError((error: Error) => {
+        // TODO: Handle error
+        console.error('Error fetching document at revision', error)
+        return [{document: null, loading: false, revisionId: revisionId}]
+      }),
+      startWith({document: null, loading: true, revisionId: revisionId}),
+    )
+}
