+++ packages/sanity/src/core/store/events/getDocumentTransactions.ts
@@ -8,6 +8,9 @@ const TRANSLOG_ENTRY_LIMIT = 50
 const documentTransactionsCache: Record<string, TransactionLogEventWithEffects[]> =
   Object.create(null)
 
+// Transactions could be cached, given they are not gonna change EVER.
+// Transactions are in an order, so if we have [rev4, rev3, rev2] and we already got [rev4, rev3] we can just get the diff between rev3 and rev2 and increment it.
+
 export async function getDocumentTransactions({
   documentId,
   client,
