--- packages/sanity/src/core/store/events/useExpandEvents.ts
@@ -1,45 +1,0 @@
-import {useMemo} from 'react'
-import {BehaviorSubject, map} from 'rxjs'
-import {type DocumentGroupEvent, isPublishDocumentVersionEvent, type SanityClient} from 'sanity'
-
-import {getDocumentTransactions} from './getDocumentTransactions'
-import {getEditEvents} from './getEditEvents'
-
-export function useExpandEvents({documentId, client}: {client: SanityClient; documentId: string}) {
-  const handler = useMemo(() => {
-    const expandedEventsMap$ = new BehaviorSubject<Map<string, DocumentGroupEvent[]>>(new Map())
-    const expandedEvents$ = expandedEventsMap$.pipe(
-      map((expandedEventsMap) => Array.from(expandedEventsMap.values()).flatMap((v) => v)),
-    )
-
-    const handleExpandEvent = async (event: DocumentGroupEvent) => {
-      if (expandedEventsMap$.getValue().has(event.id)) {
-        return // Already expanded
-      }
-      if (isPublishDocumentVersionEvent(event) && event.versionRevisionId && event.creationEvent) {
-        // This are the only events we can expand.
-        // We need to get that creation event and use versionRevisionId and fetch the transactions that occurred
-        // Since since the creation to the publish.
-        const transactions = await getDocumentTransactions({
-          client,
-          documentId,
-          fromTransaction: event.creationEvent.versionRevisionId,
-          toTransaction: event.versionRevisionId,
-        })
-        const editEvents = getEditEvents(transactions, documentId, false).map((editEvent) => ({
-          ...editEvent,
-          parentId: event.id,
-        }))
-
-        const value = expandedEventsMap$.getValue()
-        value.set(event.id, editEvents)
-        expandedEventsMap$.next(value)
-      } else {
-        console.error("This event can't be expanded", event)
-      }
-    }
-
-    return {handleExpandEvent, expandedEvents$}
-  }, [client, documentId])
-  return handler
-}
